src/test/java/com/natwest/payment/pagination/infrastructure/primary/RestPaymentsAppPageTest.java
package com.natwest.payment.pagination.infrastructure.primary;

import static org.assertj.core.api.Assertions.*;
import static com.natwest.payment.pagination.domain.PaymentsAppPagesFixture.*;

import java.util.function.Function;

import org.junit.jupiter.api.Test;

import com.natwest.payment.UnitTest;
import com.natwest.payment.JsonHelper;
import com.natwest.payment.error.domain.MissingMandatoryValueException;

@UnitTest
class RestPaymentsAppPageTest {

  @Test
  void shouldNotConvertWithoutSourcePage() {
    assertThatThrownBy(() -> RestPaymentsAppPage.from(null, source -> "test")).isExactlyInstanceOf(MissingMandatoryValueException.class);
  }

  @Test
  void shouldNotConvertWithoutMappingFunction() {
    assertThatThrownBy(() -> RestPaymentsAppPage.from(page(), null)).isExactlyInstanceOf(MissingMandatoryValueException.class);
  }

  @Test
  void shouldMapFromDomainPage() {
    RestPaymentsAppPage<String> page = RestPaymentsAppPage.from(page(), Function.identity());

    assertThat(page.getContent()).containsExactly("test");
    assertThat(page.getCurrentPage()).isEqualTo(2);
    assertThat(page.getPageSize()).isEqualTo(10);
    assertThat(page.getTotalElementsCount()).isEqualTo(21);
    assertThat(page.getPagesCount()).isEqualTo(3);
  }

  @Test
  void shouldGetPageCountForPageLimit() {
    RestPaymentsAppPage<String> page = RestPaymentsAppPage.from(pageBuilder().totalElementsCount(3).pageSize(3).build(), Function.identity());

    assertThat(page.getPagesCount()).isEqualTo(1);
  }

  @Test
  void shouldSerializeToJson() {
    assertThat(JsonHelper.writeAsString(RestPaymentsAppPage.from(page(), Function.identity()))).isEqualTo(json());
  }

  private String json() {
    return """
        {"content":["test"],\
        "currentPage":2,\
        "pageSize":10,\
        "totalElementsCount":21,\
        "pagesCount":3,\
        "hasPrevious":true,\
        "hasNext":false\
        }\
        """;
  }
}
